+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Change Log
 
-## 0.1.1 Under development
+## 0.1.1 March 7, 2024
+
+- Bug #2: Add bool typehint to `dataBsToggle()` method and add default value to `true` in  `dataBsTarget()`,
+  `dataBstoggle()`, `dataDismissTarget()`, `dataDrawerTarget()`, `dataDrawerTarget()`, `dataDropdownToggle()` and 
+  `dataToggle()` (@terabytesoftw)
 
 ## 0.1.0 March 5, 2024